@@ -1,5 +1,5 @@
 module github.com/hattya/git-utime
 
-go 1.15
+go 1.16
 
 require golang.org/x/sys v0.0.0-20210831042530-f4d43177bf5e
